Checks for Unreachable Nodes + Cycles
Added a validation step to the pipeline. This occurs after deserializing the pipeline but before running it. This validation includes first checking for cycles in the graph and using the output of that, it checks for unreachable nodes. If either check fails an error message is given with some useful information about where the check fails, and the pipeline quits.

It is also possible to perform this validation stage without running the entire pipeline. By using the "--validate_graph" command line argument the system will perform the deserialization, validation stages and optionally(if "--plot_graph" is used), followed by early exiting of the pipeline.

Commit includes a change in the example.json and warpq_pesq_dataset.json configurations. The new validation features caught an unused node that was left in there.